examples/chat app/requirements.txt
pyenet
wxpython==4.0.3
pypubsub
pywin32
git+https://code.manuelcortez.net/manuelcortez/libloader
git+https://code.manuelcortez.net/manuelcortez/platform_utils
git+https://code.manuelcortez.net/manuelcortez/accessible_output2